*: support 32-bit
This change makes the necessary code and test changes to allow
building on 32-bit, and to make all tests pass.

We also add a CI target that uses `GOARCH=386`.

Fixes #2853

+++ compaction_picker.go
@@ -1200,9 +1200,9 @@ func (p *compactionPickerByScore) pickAuto(env compactionEnv) (pc *pickedCompact
 	// bytes have been compacted further down the LSM.
 	if n := len(env.inProgressCompactions); n > 0 {
 		l0ReadAmp := p.vers.L0Sublevels.MaxDepthAfterOngoingCompactions()
-		compactionDebt := int(p.estimatedCompactionDebt(0))
+		compactionDebt := p.estimatedCompactionDebt(0)
 		ccSignal1 := n * p.opts.Experimental.L0CompactionConcurrency
-		ccSignal2 := n * p.opts.Experimental.CompactionDebtConcurrency
+		ccSignal2 := uint64(n) * p.opts.Experimental.CompactionDebtConcurrency
 		if l0ReadAmp < ccSignal1 && compactionDebt < ccSignal2 {
 			return nil
 		}
